organized folder junit tests
