+++ auth-service/src/test/java/com/webapp/enterprice/spring/auth/service/controller/AuthControllerTest.java
@@ -1,8 +1,6 @@
-package com.webapp.enterprice.spring.auth.service;
+package com.webapp.enterprice.spring.auth.service.controller;
 
-import com.webapp.enterprice.spring.auth.service.controller.AuthController;
 import com.webapp.enterprice.spring.auth.service.entity.AuthRequest;
-import com.webapp.enterprice.spring.auth.service.entity.UserRequest;
 import com.webapp.enterprice.spring.auth.service.jwt.JwtService;
 import com.webapp.enterprice.spring.auth.service.service.UserService;
 import lombok.extern.log4j.Log4j;
@@ -16,12 +14,9 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.http.MediaType;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-
-import java.util.Collections;
 
 import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
