+++ auth-service/src/test/java/com/webapp/enterprice/spring/auth/service/service/AuthServiceTest.java
@@ -1,10 +1,8 @@
-package com.webapp.enterprice.spring.auth.service;
+package com.webapp.enterprice.spring.auth.service.service;
 
 import com.webapp.enterprice.spring.auth.service.entity.User;
 import com.webapp.enterprice.spring.auth.service.entity.UserRequest;
 import com.webapp.enterprice.spring.auth.service.repository.UserRepository;
-import com.webapp.enterprice.spring.auth.service.service.UserService;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -12,7 +10,6 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-import java.util.Collections;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
