+++ eureka-service/src/main/resources/banner.txt
@@ -6,4 +6,3 @@
  | |___| |_| | | |  __/   < (_| |  ____) |  __/ |   \ V /  __/ |
  |______\__,_|_|  \___|_|\_\__,_| |_____/ \___|_|    \_/ \___|_|
 
-
