+++ order-service/src/test/java/com/webapp/enterprice/spring/order/service/service/OrderServiceTest.java
@@ -1,10 +1,9 @@
-package com.webapp.enterprice.spring.order.service;
+package com.webapp.enterprice.spring.order.service.service;
 
 import com.webapp.enterprice.spring.order.service.entity.Order;
 import com.webapp.enterprice.spring.order.service.entity.OrderRequest;
 import com.webapp.enterprice.spring.order.service.entity.Product;
 import com.webapp.enterprice.spring.order.service.repository.OrderRepository;
-import com.webapp.enterprice.spring.order.service.service.OrderService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -14,8 +13,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestTemplate;
-
-import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
